Define model variables

create_variables creates the human and mosquito variables for the model. Variables are used to track real data for each individual over time, they are read and updated by processes

The human variables are defined as:

birth - an integer representing the timestep when this individual was born

last_boosted_* - the last timestep at which this individual's immunity was boosted for tracking grace periods in the boost of immunity

is_severe - a binary indicator (0 or 1) for if the individual currently has severe malaria

ICM - Maternal immunity to clinical disease

IVM - Maternal immunity to severe disease

IB - Pre-erythoctic immunity

ICA - Acquired immunity to clinical disease

IVA - Acquired immunity to severe disease

ID - Acquired immunity to detectability

zeta - Heterogeneity of human individuals

zeta_group - Discretised heterogeneity of human individuals

variety - The variety of mosquito, either 1, 2 or 3. These are related to blood meal rate parameter

infectivity - The onward infectiousness to mosquitos

drug - The last prescribed drug

drug_time - The timestep of the last drug

create_variables(parameters)

Arguments

parameters,
model parameters created by get_parameters