Run a metapopulation model

Run a metapopulation model

run_metapop_simulation(timesteps, parameters, correlations = NULL, mixing)

Arguments

timesteps

the number of timesteps to run the simulation for (in days)

parameters

a list of model parameter lists for each population

correlations

a list of correlation parameters for each population (default: NULL)

mixing

matrix of mixing coefficients for infectivity towards mosquitoes. Rows = origin sites, columns = destinations. Each element must be between 0 and 1 and all rows must sum to 1.

Value

a list of dataframe of results