Parameterise a vaccine mass distribution strategy

distribute pre-erythrocytic vaccine to a population in an age range. Efficacy will take effect after the last dose

set_mass_pev(
  parameters,
  profile,
  timesteps,
  coverages,
  min_ages,
  max_ages,
  min_wait,
  booster_spacing,
  booster_coverage,
  booster_profile
)

Arguments

parameters

a list of parameters to modify

profile

a list of details for the vaccine profile, create with create_pev_profile

timesteps

a vector of timesteps for each round of vaccinations

coverages

the coverage for each round of vaccinations

min_ages

for the target population, inclusive (in timesteps)

max_ages

for the target population, inclusive (in timesteps)

min_wait

the minimum acceptable time since the last vaccination (in timesteps); When using both set_mass_pev and set_pev_epi, this represents the minimum time between an individual being vaccinated under one scheme and vaccinated under another.

booster_spacing

the timesteps (following the final primary dose) at which booster vaccinations are administered

booster_coverage

a matrix of coverages (timesteps x boosters) specifying the proportion the previously vaccinated population to continue receiving booster doses. The rows of the matrix must be the same size as timesteps. The columns of the matrix must be the same size as booster_spacing.

booster_profile

list of lists representing each booster profile, the outer list must be the same length as booster_spacing. Create vaccine profiles with create_pev_profile