Parameterise a pre-erythrocytic vaccine with an EPI strategy

distribute vaccine when an individual becomes a certain age. Efficacy will take effect after the last dose

set_pev_epi(
  parameters,
  profile,
  coverages,
  timesteps,
  age,
  min_wait,
  booster_timestep,
  booster_coverage,
  booster_profile,
  seasonal_boosters = FALSE
)

Arguments

parameters

a list of parameters to modify

profile

primary vaccine profile of type PEVProfile

coverages

a vector of coverages for the primary doses

timesteps

a vector of timesteps associated with coverages

age

the age when an individual will receive the first dose of the vaccine (in timesteps)

min_wait

the minimum acceptable time since the last vaccination (in timesteps); When seasonal_boosters = TRUE, this represents the minimum time between an individual receiving the final dose and the first booster. When using both set_mass_pev and set_pev_epi, this represents the minimum time between an individual being vaccinated under one scheme and vaccinated under another.

booster_timestep

the timesteps (following the final dose) at which booster vaccinations are administered

booster_coverage

the proportion of the vaccinated population relative to the last vaccination (whether a previous booster or the primary series)

booster_profile

list of booster vaccine profiles, of type PEVProfile, for each timestep in booster_timeteps

seasonal_boosters

logical, if TRUE the first booster timestep is relative to the start of the year, otherwise they are relative to the last dose