Get model parameters

get_paramaters creates a named list of parameters for use in the model. These parameters are passed to process functions. These parameters are explained in "The US President's Malaria Initiative, Plasmodium falciparum transmission and mortality: A modelling study."

NOTE: this function is likely to be extended to read in command line / config file parameters

The parameters are defined as:

days_per_timestep - the number of days to model per timestep

rd - the rate at which humans move from state D to A

ra - the rate at which humans move from state A to U

ru - the rate at which humans move from state U to S

rel - the rate at which mosquitos move from state E to L

rl - the rate at which mosquitos move from state L to P

rpl - the rate at which mosquitos move from state P to Sm

mup - the rate at which pupal mosquitos die

mum - the rate at which developed mosquitos die

beta - the average number of eggs laid per female mosquito per day

human population - the number of humans to model

mosquito limit - the maximum number of mosquitos to allow for in the

days_per_timestep - the number of days to model per timestep

get_parameters()