Get model parameters

get_paramaters creates a named list of parameters for use in the model. These parameters are passed to process functions. These parameters are explained in "The US President's Malaria Initiative, Plasmodium falciparum transmission and mortality: A modelling study."

NOTE: this function is likely to be extended to read in command line / config file parameters

The parameters are defined below.

fixed state transitions:

dd - the delay for humans to move from state D to A

da - the delay for humans to move from state A to U

du - the delay for humans to move from state U to S

del - the delay for mosquitos to move from state E to L

dl - the delay for mosquitos to move from state L to P

dpl - the delay mosquitos to move from state P to Sm

mup - the rate at which pupal mosquitos die

mum - the rate at which developed mosquitos die

immunity decay rates:

rm - decay rate for maternal immunity to clinical disease

rvm - decay rate for maternal immunity to severe disease

rb - decay rate for acquired pre-erytrhrocytic immunity

rc - decay rate for acquired immunity to clinical disease

rva - decay rate for acquired immunity to severe disease

rid - decay rate for acquired immunity to detectability

probability of pre-erythrocytic infection:

b0 - maximum probability due to no immunity

b1 - maximum reduction due to immunity

ib0 - scale parameter

kb - shape parameter

probability of clinical infection:

phi0 - maximum probability due to no immunity

phi1 - maximum reduction due to immunity

ic0 - scale parameter

kc - shape parameter

probability of severe infection:

severe_enabled - whether to model severe disease

theta0 - maximum probability due to no immunity

theta1 - maximum reduction due to immunity

iv0 - scale parameter

kv - shape parameter

fv0 - age dependent modifier

av - age dependent modifier

gammav - age dependent modifier

immunity reducing probability of detection:

fd0 - time-scale at which immunity changes with age

ad - scale parameter relating age to immunity

gammad - shape parameter relating age to immunity

d1 - minimum probability due to immunity

dmin - minimum probability due to immunity NOTE: there appears to be a mistake here!

id0 - scale parameter

kd - shape parameter

immunity boost grace periods:

ub - period in which pre-erythrocytic immunity cannot be boosted

uc - period in which clinical immunity cannot be boosted

uv - period in which severe immunity cannot be boosted

ud - period in which immunity to detectability cannot be boosted

blood meal rates:

av1 - blood meal rate for the first variety of mosquitos

av2 - blood meal rate for the second variety of mosquitos

av3 - blood meal rate for the third variety of mosquitos

infectivity towards mosquitos:

cd - infectivity of clinically diseased humans towards mosquitos

gamma1 - parameter for infectivity of asymptomatic humans

cu - infectivity of sub-patent infection

unique biting rate:

a0 - age dependent biting parameter

rho - age dependent biting parameter

sigma_squared - heterogeneity parameter

n_heterogeneity_groups - number discretised groups for heterogeneity, used for sampling mothers

mortality parameters:

mortality_probability_table - a vector of mortality rates for humans between the ages of 0 and 100

v - mortality scaling factor from severe disease

pcm - new-born clinical immunity relative to mother's

pvm - new-born severe immunity relative to mother's

me - early stage larval mortality rate

ml - late stage larval mortality rate

carrying capacity parameters:

model_seasonality - boolean switch TRUE iff the simulation models seasonal rainfall

K0 - carrying capacity

g0 to g3 - rainfall shape parameters

h1 to h3 - rainfall shape parameters

gamma - effect of density dependence on late instars relative to early instars

initial state proportions:

s_proportion - the proportion of human_population that begin as Susceptable

d_proportion - the proportion of human_population that begin with clinical disease

a_proportion - the proportion of human_population that begin as Asymptomatic

u_proportion - the proportion of human_population that begin as Subpatents

initial immunity values:

init_icm - the immunity from clinical disease at birth

init_ivm - the immunity from severe disease at birth

init_ib - the initial pre-erythrocitic immunity

init_ica - the initial acquired immunity from clinical disease

init_iva - the initial acquired immunity from severe disease

init_id - the initial acquired immunity to detectability

miscellaneous:

de - delay for infection

dem - delay for infection in mosquitoes

beta - the average number of eggs laid per female mosquito per day

human_population - the number of humans to model

mosquito_limit - the maximum number of mosquitos to allow for in the simulation

density - the initial number of mosquitos per human

days_per_timestep - the number of days to model per timestep

get_parameters(overrides = list())