Get model parameters

get_paramaters creates a named list of parameters for use in the model. These parameters are passed to process functions. These parameters are explained in "The US President's Malaria Initiative, Plasmodium falciparum transmission and mortality: A modelling study."

get_parameters(overrides = list())

Arguments

overrides
a named list of parameter values to use instead of defaults The parameters are defined below.

fixed state transitions:

dd - the delay for humans to move from state D to A

dt - the delay for humans to move from state Tr to Ph

da - the delay for humans to move from state A to U

du - the delay for humans to move from state U to S

del - the delay for mosquitos to move from state E to L

dl - the delay for mosquitos to move from state L to P

dpl - the delay mosquitos to move from state P to Sm

mup - the rate at which pupal mosquitos die

mum - the rate at which developed mosquitos die

immunity decay rates:

rm - decay rate for maternal immunity to clinical disease

rvm - decay rate for maternal immunity to severe disease

rb - decay rate for acquired pre-erytrhrocytic immunity

rc - decay rate for acquired immunity to clinical disease

rva - decay rate for acquired immunity to severe disease

rid - decay rate for acquired immunity to detectability

probability of pre-erythrocytic infection:

b0 - maximum probability due to no immunity

b1 - maximum reduction due to immunity

ib0 - scale parameter

kb - shape parameter

probability of clinical infection:

phi0 - maximum probability due to no immunity

phi1 - maximum reduction due to immunity

ic0 - scale parameter

kc - shape parameter

probability of severe infection:

severe_enabled - whether to model severe disease

theta0 - maximum probability due to no immunity

theta1 - maximum reduction due to immunity

iv0 - scale parameter

kv - shape parameter

fv0 - age dependent modifier

fvt - reduced probability of death due to treatment

av - age dependent modifier

gammav - age dependent modifier

immunity reducing probability of detection:

fd0 - time-scale at which immunity changes with age

ad - scale parameter relating age to immunity

gammad - shape parameter relating age to immunity

d1 - minimum probability due to immunity

id0 - scale parameter

kd - shape parameter

immunity boost grace periods:

ub - period in which pre-erythrocytic immunity cannot be boosted

uc - period in which clinical immunity cannot be boosted

uv - period in which severe immunity cannot be boosted

ud - period in which immunity to detectability cannot be boosted

infectivity towards mosquitos:

cd - infectivity of clinically diseased humans towards mosquitos

gamma1 - parameter for infectivity of asymptomatic humans

cu - infectivity of sub-patent infection

ct - infectivity of treated infection

unique biting rate:

a0 - age dependent biting parameter

rho - age dependent biting parameter

sigma_squared - heterogeneity parameter

n_heterogeneity_groups - number discretised groups for heterogeneity, used for sampling mothers

mortality parameters:

average_age - the average age of humans (in timesteps)

v - mortality scaling factor from severe disease

pcm - new-born clinical immunity relative to mother's

pvm - new-born severe immunity relative to mother's

me - early stage larval mortality rate

ml - late stage larval mortality rate

carrying capacity parameters:

model_seasonality - boolean switch TRUE iff the simulation models seasonal rainfall

g0 - rainfall fourier parameter

g - rainfall fourier parameter

h - rainfall fourier parameters

gamma - effect of density dependence on late instars relative to early instars

initial state proportions:

s_proportion - the proportion of human_population that begin as Susceptable

d_proportion - the proportion of human_population that begin with clinical disease

a_proportion - the proportion of human_population that begin as Asymptomatic

u_proportion - the proportion of human_population that begin as subpatents

t_proportion - the proportion of human_population that begin treated

initial immunity values:

init_icm - the immunity from clinical disease at birth

init_ivm - the immunity from severe disease at birth

init_ib - the initial pre-erythrocitic immunity

init_ica - the initial acquired immunity from clinical disease

init_iva - the initial acquired immunity from severe disease

init_id - the initial acquired immunity to detectability

incubation periods:

de - delay for infection

dem - delay for infection in mosquitoes

vector biology: species specific values are vectors

beta - the average number of eggs laid per female mosquito per day

total_M - the initial number of adult mosquitos in the simulation

init_foim - the FOIM used to calculate the equilibrium state for mosquitoes

species - names of the species in the simulation

species_proportions - the relative proportions of each species

blood_meal_rates - the blood meal rates for each species

Q0 - proportion of blood meals taken on humans

foraging_time - time spent taking blood meals

feeding cycle:

bednets - boolean for if bednets are enabled

rn - probability mosquito is repelled by the bednet

rnm - minimum probability mosquito is repelled by the bednet

dn0 - probability killed by the bednet

spraying - boolean for if indoor spraying is enabled

rs - probability repelled by indoor spraying

phi_indoors - proportion of bites taken indoors

phi_bednets - proportion of bites taken in bed

endophily - proportion of mosquitoes resting indoors after feeding with no intervention

gammas - the half-life of spraying efficacy (timesteps)

gamman - the half-life of bednet efficacy (timesteps)

please set vector control strategies using set_betnets and set_spraying

treatment parameters: I recommend setting these with the convenience functions in drug_parameters.R

drug_efficacy - a vector of efficacies for available drugs

drug_rel_c - a vector of relative onwards infectiousness values for drugs

drug_prophylaxis_shape - a vector of shape parameters for weibull curves to model prophylaxis for each drug

drug_prophylaxis_scale - a vector of scale parameters for weibull curves to model prophylaxis for each drug

ft - probability of seeking treatment if clinically diseased

clinical_treatment_drugs - a vector of drugs that are avaliable for clinically diseased (these values refer to the index in drug_* parameters)

clinical_treatment_coverage - a vector of coverage values for each drug

RTS,S paramters:

rtss_vmax - the maximum efficacy of the vaccine

rtss_alpha - shape parameter for the vaccine efficacy model

rtss_beta - scale parameter for the vaccine efficacy model

rtss_cs - peak parameters for the antibody model (mean and std. dev)

rtss_cs_boost - peak parameters for the antibody model for booster rounds (mean and std. dev)

rtss_rho - delay parameters for the antibody model (mean and std. dev)

rtss_rho_boost - delay parameters for the antibody model for booster rounds (mean and std. dev)

rtss_ds - delay parameters for the antibody model (mean and std. dev)

rtss_dl - delay parameters for the antibody model (mean and std. dev)

I recommend setting strategies with the convenience functions in vaccine_parameters.R:set_rtss

MDA and SMC parameters: We recommend setting these with convenience functions in mda_parameters.R

TBV parameters:

tbv_mt - effect on treated infectiousness

tbv_md - effect on diseased infectiousness

tbv_ma - effect on asymptomatic infectiousness

tbv_mu - effect on subpatent infectiousness

tbv_k - scale parameter for effect on infectiousness

tbv_tau - peak antibody parameter

tbv_rho - antibody component parameter

tbv_ds - antibody short-term delay parameter

tbv_dl - antibody long-term delay parameter

tbv_tra_mu - transmission reduction parameter

tbv_gamma1 - transmission reduction parameter

tbv_gamma2 - transmission reduction parameter

I recommend setting tbv strategies with the convenience functions in vaccine_parameters.R, these are the same as for RTS,S

rendering: All values are in timesteps and all ranges are inclusive

prevalence_rendering_min_ages - the minimum ages for clinical prevalence outputs

prevalence_rendering_max_ages - the corresponding max ages

incidence_rendering_min_ages - the minimum ages for clinical incidence outputs

incidence_rendering_max_ages - the corresponding max ages

severe_prevalence_rendering_min_ages - the minimum ages for severe prevalence outputs

severe_prevalence_rendering_max_ages - the corresponding max ages

severe_incidence_rendering_min_ages - the minimum ages for severe incidence outputs

severe_incidence_rendering_max_ages - the corresponding max ages

miscellaneous:

human_population - the number of humans to model

mosquito_limit - the maximum number of mosquitos to allow for in the simulation

individual_mosquitoes - boolean whether adult mosquitoes are modelled individually or compartmentaly