Parameterise an indoor spraying strategy

The model will apply indoor spraying at timesteps to a random sample of the entire human population. The sample size will be a proportion of the human population taken from the corresponding coverages. The sample can contain humans who have already benefited from spraying.

If a human in the sample lives in a sprayed house, the efficacy of the spraying will be returned to the maximum.

The structure for the indoor residual spraying model is documented in the S.I. of 10.1038/s41467-018-07357-w

set_spraying(
  parameters,
  timesteps,
  coverages,
  ls_theta,
  ls_gamma,
  ks_theta,
  ks_gamma,
  ms_theta,
  ms_gamma
)

Arguments

parameters

a list of parameters to modify

timesteps

the timesteps at which to spray

coverages

the proportion of the population who get indoor spraying

ls_theta

matrix of mortality parameters With nrows=length(timesteps), ncols=length(species)

ls_gamma

matrix of mortality parameters per timestep With nrows=length(timesteps), ncols=length(species)

ks_theta

matrix of feeding success parameters per timestep With nrows=length(timesteps), ncols=length(species)

ks_gamma

matrix of feeding success parameters per timestep With nrows=length(timesteps), ncols=length(species)

ms_theta

matrix of deterrence parameters per timestep With nrows=length(timesteps), ncols=length(species)

ms_gamma

matrix of deterrence parameters per timestep With nrows=length(timesteps), ncols=length(species)